// frontend/src/app/core/models/user.model.ts
export interface User {
  id: number;
  employeeId: string;
  name: string;
  email: string;
  role: 'EMPLOYEE' | 'MANAGER' | 'ADMIN';
  department: string;
  designation: string;
  managerId?: number;
  managerName?: string;
  phone?: string;
  address?: string;
  joiningDate?: string;
  salary?: number;
  status?: 'ACTIVE' | 'INACTIVE';
}

export interface AuthResponse {
  token: string;
  user: User;
}